requirements.txt
aiohttp==3.12.15
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
asyncio-throttle==1.0.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
google-auth==2.40.3
google-genai==1.32.0
google-generativeai==0.8.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.4
openai==1.102.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pyyaml==6.0.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
websockets==15.0.1
